requirement.txt
asgiref==3.8.1
Django==5.1.6
sqlparse==0.5.3
requests==2.32.3
djangorestframework==3.15.2
pillow==11.1.0
dj-rest-auth==7.0.1